RF-35 refactor firebase auth using v9 functions
