--- src/context/FirebaseContext.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const FirebaseContext = createContext(null);
-
-export default FirebaseContext;
