+++ src/hooks/useAuthListener.js
@@ -1,14 +1,14 @@
-import { useContext, useState, useEffect } from 'react';
-import FirebaseContext from '../context/FirebaseContext';
+import { useState, useEffect } from 'react';
+import { onAuthStateChanged } from 'firebase/auth';
+import { auth } from '../services/firebase';
 
 function useAuthListener() {
-    const auth = JSON.parse(localStorage.getItem('authUser'));
+    const savedUser = JSON.parse(localStorage.getItem('authUser'));
 
-    const [user, setUser] = useState(auth);
-    const { firebase } = useContext(FirebaseContext);
+    const [user, setUser] = useState(savedUser);
 
     useEffect(() => {
-        const listener = firebase.auth().onAuthStateChanged(authUser => {
+        const listener = onAuthStateChanged(auth, authUser => {
             if (authUser) {
                 localStorage.setItem('authUser', JSON.stringify(authUser));
                 setUser(authUser);
